add positive nbr

+++ 0x0A-argc_argv/3-mul.c
@@ -6,7 +6,7 @@
  * Return: void
  */
 
-void	print_number(long n)
+void	print_number(int n)
 {
 	if (n < 0)
 	{
